Rename `Patch` to `PatchGrid`.
This change is made so that we can have a separate `Patch` class that
contains the state used for the run, i.e. there will be a `PatchGrid`
class with just the coordinate/size data, and a `Patch` class that
contains both a `PatchGrid` and the data used to solve a particular
problem.